(maint) never fall back to standard strategy
This alters the nodelay provisioner strategy to always indicate
it is done with provisioning at the round it completes and not
fall back to the standard provisioner. Since the provisioning
loop can execute very slowly, this will keep the provisioning
loops tighter to incrementally provide nodes instead of blocking
until all the nodes are set up.